refactor v2 lookup arg resolution
